Added: Client bricks explosion
The bricks explode when they get hit by the ball or the level changes.
The explosion consists of an animated cursor that is drawn on a thread in the middle of the bricks position, with a pause of 6 Jiffies / 100 ms between each cursors frame.
Added variable description and client restriction to only allow clients that are playing the game to move.